backend/requirements.txt
asgiref==3.6.0
boto3==1.26.106
botocore==1.29.106
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jmespath==1.0.1
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
